Add files via upload
We have named this project as Job Openings because in this project we have done job analysis to find out how many jobs are there in which city, which company is offering how many jobs and which field has the most jobs. All these things have been analyzed in this project.